+++ src/main/java/dk/sdu/swe/domain/models/Company.java
@@ -1,9 +1,6 @@
 package dk.sdu.swe.domain.models;
 
-import dk.sdu.swe.data.converters.CompanyDetailsConverter;
-
 import javax.persistence.*;
-import java.util.Collection;
 import java.util.List;
 
 @Entity
@@ -12,7 +9,7 @@ public class Company {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
 
@@ -38,11 +35,11 @@ public class Company {
     public Company() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
